added color to GUI

+++ src/more_vectors.hpp
@@ -10,7 +10,7 @@ struct vector4 {
     vector4(float x_, float y_, float z_, float q_)
         : x(x_), y(y_), z(z_), q(q_) {}
 
-    float x, y, z, q;
+    uint32_t x, y, z, q;
 };
 
 }  // namespace MoreVectors
